Add usage example on replit to README
In an effort to provide a more user-friendly and accessible experience, an example usage on the platform 'replit' was added to the README file. By using this online platform, users who don't want to install the software can test it directly. An image and link to the replit were included. Accompanying explanations were updated accordingly to facilitate better understanding.